font-patcher: Fix monospaced glyph collection detection
[why]
When only one symbol glyph is examined we conclude that it comes from a
monospaced glyph set.

This might be correct or not, but when we can not positively say it is
monospaced we should not handle it as monospaced.

[how]
We require at least TWO glyphs with the same width (and no glyph
with a different width) until we set the 'advance' bounding box
property. Which says that this particular glyph subset is monospaced.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>