font-patcher: Special rule for SymbolsOnly font
[why]
When patching the SymbolsOnly font via gotta-patch-em-all we create a
'Nerd Font Propo'. With the previous change that will be reflected by a
changed name.

The name change can come unexpected for users.

And in fact a 'Nerd Font' version would be ridiculous as there are no
preexsiting glyphs in the font.

So we should keep the previous naming:
* `Symbols Only Nerd Font` (which is in fact a Nerd Font Propo)
* `Symbols Only Nerd Font Mono` (for the monospaced version)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>