Update sankey diagram
[why]
We added a lot of glyphs in Seti + Custom and the new range Material
Design Icons.

[how]
Create new graphics with generator, see README.md.
Use Inkscape to implant into old image.

Note:
In the totals just the new Material Design Icons are included, the old
(obsolete, to be removed) range is not counted at all anymore.

Also did not change the distribution of the diagram. When we would
create a real sankey diagram (with the actual numbers) all the other
sets would be dwarfted by the new Material Design Icons.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>