Mark old MDI as obsolete on cheat sheet
[why]
The (old) Material Design Icons are to be removed.
We should communicate that on the Cheat Sheet.

[how]
See commit
  4452ceee5  Add possibility to add "obsolete" to glyphs

that implements the CSS stuff to display an 'obsolete' marker.

Change the cheat-sheet generatot to actually insert the markers for the
appropriate codepoints. That is, the codepoints are not checked but the
ID.

Fixes: #1096

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>