Publish contributors after allcontributors
[why]
After adding a contributor with allcontributors the website is not
updated.

[how]
Obviously there is a setting if a dependant workflow should be started.
Set that value to 'true'.

Also fix commented-out section of workflow triggers for 'release', that
should not be triggered by allcontributors, when that is commented-in
again (if ever).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>