Drop support for Python 2
[why]
Python 2 is long since EOL.

With the last commit we want to use Enums, which are not available in
Python 2.

I believe Python 2 broke some time before already, I stopped caring for
Python 2 some time ago.

This does not change any existing code. It just documents that there are
no efforts anymore to support Python 2.

This is PR #1121

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>