test-powerline: Enable use with old bash
[why]
On MacOS Ventura the unicode codepoints are shown instead of the actual
glyphs.

[how]
On that OS the bash is rather old and does not have the unicode escape
sequence \u. As it is rather hard to convert it to octal UTF8 we instead
embed simply the concrete characters, as done in the i_*.sh scripts.

Note: zsh on that OS does have it, though.

[note]
I'm no Mac expert, just trying to get this working.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>