font-patcher: Make name shortening optional
[why]
The option --makegroups shortens the name suffix and the styles
rigorously. That is needed for fonts with long names and/or long style
descriptions.
But some simple font would not need that. With the option we can adjust
the amount of shortening to fit the specific font (family).

Unfortunately we can not automatically detect that, because the
font-patcher script can not assume anything about a complete family as
every font file is patched individually.

[how]
Add optional numeric parameter to option '--makegroups'.
The values can be:
  0 - turned off, use old naming scheme
  1 - turned on, shortening                         'Bold' to 'Bd'
  2 - turned on, shortening 'Nerd Font' to 'NF'
  3 - turned on, shortening 'Nerd Font' to 'NF' and 'Bold' to 'Bd'
  4 - turned on, no shortening

If --makegroups is not specified this is the same as '--makegroups 0'.
If --makegroups is specified without number that means '--makegroups 3'.

[note]
The argparser module takes the last specified parameter as the one to
use: If you specify --makegroups multiple times only the last one will
be used.

This can be used for gotta-patch-em-all. Per default it (will) specify
--makegroups 3. But via config.cfg we can specify a font centric other
number for individual fonts. The config.cfg values are added later in
the command line, so they take precedence.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>