Revert "font-patcher: Use WIN metrics in all conflicting cases"
This reverts commit 2dd2e8113b97059db68e09fdc523b4fc8e7a04d9,
and adapts the code to more recent changes (logging, enums).

[why]
Lekton has a too wide line spacing.

Lekton was the only font that selected TYPO but we forced it to WIN,
because not all glyphs fit into the values from TYPO.
But that seems to be wrong. Examining the glyphs that are really bigger
than the TYPO line spaces, these are only graphical glyphs that shall
span multiple lines. So I guess we should revert that change and render
Lekton with the TYPO values.

[note]
https://github.com/ryanoasis/nerd-fonts/issues/1056#issuecomment-1529141863