test-fonts: Fix missing codepoints and allow MacOS
[why]
If a set has multiple ranges defined only the last of that ranges is
actually printed.

[how]
With commit
  28780da9f  Fixes build errors (ShellCheck)

a false positive of spellcheck lead to a 'correction' in the code that
actually broke it. The mapfile does not accumulate the sequences but
fills it in and so just the last sequence 'survives'.

Dropping mapfile also enables this script on MacOS, as that ancient
bash does not have mapfile.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>